Added progress bars to temperature
I’m not really happy with the way this works. The progress bar remains
at the end of the run and has weird artifacts after exiting. It
displays an incorrect percentage after exiting and will sometimes screw
up the table render (but not consistently).  It would be nice to add
network stats as well.

This also contains a fix for correctly referencing the vapor host url
and sets a default.